pyproject.toml
[project]
name = "hydran"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.12",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
]
